Update Prog21.java
int型変数のyを doubleにキャスト（強制型変換）した

+++ src/main/java/Prog21.java
@@ -6,6 +6,6 @@ public class Prog21 {
         String sikiB = "y/x = ";
         System.out.println("x=" + 2 + ",y=" + 7 +"とすると、");
         System.out.println(sikiA + x+y);
-        System.out.println(sikiB + y/x);
+        System.out.println(sikiB + (double)y/x);
     }
 }
